Restaurants modifications
Done to extract location even before a cookie is set, so the "Show me
Hamilton" will adapt and show the locale a person is located. We still
need to script it to take full advantage of the cookies, although it was
being implemented already. The code uses ipinfo.io, but once PECL is
available and geoip is installed, it is ready to use PHP's native
support.  The modifications also are working to adjust the search
between miles an kilometers, although it is not fully implemented
everywhere yet.